Task3/finalpro/requirements.txt
feedparser
newspaper3k
requests
pymongo
redis
langchain
langchain-community
sentence-transformers
faiss-cpu
streamlit
pandas
plotly
python-dotenv
beautifulsoup4
fastapi
uvicorn
